Update cached role_map if AWS gives AccessDenied
AWS returning AccessDenied is a good indicator that our cached role map isn't accurate.
This makes the tool refresh the role map if an AccessDenied is encountered
This also reimplements #156 differently.
* We shouldn't test if the role_arn is present in the role_map in a function called `get_role_map`
* We don't want to check if the role_arn is present in cases where we merely want to refresh the role map

This also sends a message to the console stderr if the user provides an invalid role ARN as
a command line argument, before spawning the role picker.
* Ideally this message would be shown in the role picker but it looks like it would take some
  heavy lifting to enable triggering a `setMessage` in index.js from the beginning of a
  redirect flow such that it shows up after auth completes and the user is at the role picker