Update dev ID token for role Makefile to use dev IdP Improve error messaging for uncaught exceptions Change from outputting export to rendering shell script and add shell prompt info
This changes how the environment variables that maws outputs are conveyed to the parent shell.
Previously a single command was emitted (for non Windows clients) in the form of
"export foo=bar baz=buz"

This is limited to a single command because of how this output is being interpreted by the $(maws)
subshell wrapper.

To work around this limitation maws now produces a temporary shell script which is "sourced" into
the parent shell. This way an arbitrary number of commands can be injected into the parent shell

The temporary shell script
* exports environment variables
* unsets environment variables that should no longer be set
* provisions a function that outputs the current profile name
* modifies the $PS1 prompt to call this function
* deletes itself

This accomplishes two things
* Ensures that a user never ends up with a conflicting set of environment variables (like
  AWS_PROFILE and AWS_ACCESS_KEY_ID) from using one output setting in a given terminal, then using
  a different output setting
* Injects the current profile name into the command prompt so it's clear which account and role
  the user is operating as